feat: Simplify import paths by consolidating type imports from a single module
