+++ server/src/models/UsersUtils/users.methods.ts
@@ -2,7 +2,7 @@ import jwt from "jsonwebtoken";
 import bcrypt from "bcryptjs";
 import { UserModel } from "./../users.model";
 import { UserRequestObject } from "../../types/UserRequest";
-import { CustomClaims } from "../../types/JwtSignClaims";
+import { CustomClaims } from "../../types";
 import { CustomError } from "../../errors/CustomError";
 
 export function createJWTMethod(this: UserModel): string {
